requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioquic==1.2.0
aiosignal==1.3.1
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
Django==5.1.2
dnspython==2.7.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
multidict==6.1.0
postfix-mta-sts-resolver==1.4.0
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.4.0
pycparser==2.22
pylsqpack==0.3.18
pyOpenSSL==24.2.1
PyYAML==6.0.2
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.2.3
yarl==1.14.0
